pyproject.toml
[project]
name = "test-fastapi-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Layss kheir",email = "laysskheir@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "openai (>=1.97.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
